requirements.txt
# Core dependencies for RAG Pipeline

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# LLM Client
groq==0.11.0

# Vector Store (Supabase with pgvector)
supabase==2.10.0
httpx==0.27.2

# Document Processing
python-docx==1.1.0

# Embeddings
google-generativeai>=0.5.0
sentence-transformers==2.3.1
torch>=2.0.0

# Configuration
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
numpy>=1.24.0
